requirements.txt
decorator==4.0.10
defusedxml==0.4.1
Django==1.10.2
django-allauth==0.27.0
django-model-utils==2.6
factory-boy==2.7.0
fake-factory==0.7.2
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
lxml==3.6.4
oauthlib==2.0.0
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.8
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.5.3
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.7.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
